# docs/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Union Bank API",
    "description": "REST API for the Union Bank Management System.\n\n**API Versions**\n- `/api/v1/` \u2014 Legacy endpoints (bare response models, backward compatible)\n- `/api/v2/` \u2014 Current endpoints with standardised `ApiResponse[T]` envelope\n\nAll endpoints return JSON. Authentication uses Bearer JWT tokens.\nUse `/docs` for interactive API documentation.",
    "termsOfService": "https://union-bank.example.com/terms",
    "contact": {
      "name": "Union Bank Dev Team"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "2.0.0"
  },
  "paths": {
    "/api/v2/auth/login": {
      "post": {
        "summary": "V2 Customer Login",
        "description": "Authenticate a customer and return a JWT access + refresh token pair.\n\nTokens are set as httpOnly cookies (primary) and returned in the\nresponse body (backward compatibility).",
        "operationId": "v2_customer_login_api_v2_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_TokenData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/auth/register": {
      "post": {
        "summary": "V2 Customer Register",
        "description": "Register a new customer account.",
        "operationId": "v2_customer_register_api_v2_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MessageData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/auth/admin-login": {
      "post": {
        "summary": "V2 Admin Login",
        "description": "Authenticate as admin and return a JWT access + refresh token pair.\n\nTokens are set as httpOnly cookies (primary) and returned in the\nresponse body (backward compatibility).",
        "operationId": "v2_admin_login_api_v2_auth_admin_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/unionbank__entrypoints__api__models__AdminLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_TokenData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/auth/refresh": {
      "post": {
        "summary": "V2 Refresh Token",
        "description": "Exchange a refresh token for a new access + refresh token pair.\n\nAccepts the refresh token from either:\n1. The request body (backward compatibility)\n2. The ub_refresh_token httpOnly cookie (new cookie-based flow)\n\nThe previous refresh token is revoked (rotation) so it cannot be reused.",
        "operationId": "v2_refresh_token_api_v2_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/unionbank__entrypoints__api__models__RefreshRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_TokenData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/login": {
      "post": {
        "summary": "Customer Login",
        "description": "Authenticate a customer and return a JWT access + refresh token pair.\n\nTokens are set as httpOnly cookies (Secure, SameSite=Strict) and also\nreturned in the response body for backward compatibility with Bearer\ntoken clients.",
        "operationId": "customer_login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "deprecated": true
      }
    },
    "/api/auth/register": {
      "post": {
        "summary": "Customer Register",
        "description": "Register a new customer account.",
        "operationId": "customer_register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/admin-login": {
      "post": {
        "summary": "Admin Login",
        "description": "Authenticate as admin and return a JWT access + refresh token pair.\n\nTokens are set as httpOnly cookies (Secure, SameSite=Strict) and also\nreturned in the response body for backward compatibility.",
        "operationId": "admin_login_api_auth_admin_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/unionbank__entrypoints__api__main__AdminLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/refresh": {
      "post": {
        "summary": "Refresh Token",
        "description": "Exchange a refresh token for a new access + refresh token pair.\n\nAccepts the refresh token from either:\n1. The request body (backward compatibility with Bearer token clients)\n2. The ub_refresh_token httpOnly cookie (new cookie-based flow)\n\nThe previous refresh token is revoked (rotation) so it cannot be reused.",
        "operationId": "refresh_token_api_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/unionbank__entrypoints__api__main__RefreshRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/2fa/status": {
      "get": {
        "summary": "Admin Totp Status",
        "description": "Check if 2FA is enabled for the current admin.",
        "operationId": "admin_totp_status_api_admin_2fa_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TOTPStatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/2fa/setup": {
      "get": {
        "summary": "Admin Totp Setup",
        "description": "Generate a new TOTP secret and provisioning URI for the admin user.",
        "operationId": "admin_totp_setup_api_admin_2fa_setup_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TOTPSetupResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/2fa/verify": {
      "post": {
        "summary": "Admin Totp Verify",
        "description": "Verify a TOTP code to enable 2FA for the admin account.",
        "operationId": "admin_totp_verify_api_admin_2fa_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TOTPVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/2fa/disable": {
      "post": {
        "summary": "Admin Totp Disable",
        "description": "Disable 2FA for the admin account (requires current TOTP code).",
        "operationId": "admin_totp_disable_api_admin_2fa_disable_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TOTPVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_MessageData_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MessageData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "title": "ApiResponse[MessageData]"
      },
...
      "ApiResponse_TokenData_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TokenData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "title": "ApiResponse[TokenData]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "account_number": {
            "type": "string",
            "title": "Account Number",
            "description": "10-digit account number"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "title": "Password",
            "description": "Account password"
          }
        },
        "type": "object",
        "required": [
          "account_number",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "success"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse"
      },
...
      "RegisterRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Full name (2-50 chars, letters/spaces only)"
          },
          "age": {
            "type": "integer",
            "maximum": 120.0,
            "minimum": 18.0,
            "title": "Age",
            "description": "Age (18-120)"
          },
          "gender": {
            "type": "string",
            "title": "Gender",
            "description": "Gender"
          },
          "mobile": {
            "type": "string",
            "title": "Mobile",
            "description": "10-digit mobile number starting with 6-9"
          },
          "email": {
            "type": "string",
            "title": "Email",
            "description": "Valid email address"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password",
            "description": "Password: min 8 chars, upper+lower+digit"
          },
          "confirm_password": {
            "type": "string",
            "title": "Confirm Password",
            "description": "Must match password"
          }
        },
        "type": "object",
        "required": [
          "name",
          "age",
          "gender",
          "mobile",
          "email",
          "password",
          "confirm_password"
        ],
        "title": "RegisterRequest"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "expires_in": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires In"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "role"
        ],
        "title": "TokenResponse"
      },
...
      "unionbank__entrypoints__api__main__AdminLoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username",
            "description": "Admin username"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "title": "Password",
            "description": "Admin password"
          },
          "totp_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 6,
                "minLength": 6
              },
              {
                "type": "null"
              }
            ],
            "title": "Totp Code",
            "description": "TOTP code (required if 2FA is enabled)"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "AdminLoginRequest"
      },
...
      "unionbank__entrypoints__api__models__AdminLoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username",
            "description": "Admin username"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "title": "Password",
            "description": "Admin password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "AdminLoginRequest"
      },
      "unionbank__entrypoints__api__models__RefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token",
            "description": "Refresh token from previous login"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}